shell: move ErrorWaitGroup to weed/util
